Alter install.sh and uninstall.sh.
1. Install .vim and .vimrc to ~ by create symbol link.
2. Before Install, install.sh will back up the exist .vim and .vimrc by mv opertion.
3. uninstall.sh just rm the ~/.vim and ~/.vimrc , will interaction.

+++ uninstall.sh
@@ -1,26 +1,31 @@
-#!/usr/bin/env sh
+#!/bin/bash
 
-#define path 
+#define path
 PLUGIN_INSTALL_DIR=$HOME/.vim
 CONFIG_INSTALL_DIR=$HOME/.vimrc
-YCM_CONFIG_INSTALL_DIR=$HOME/.ycm_extra_conf.py
 
-#delete .vim
-if [ -d "$PLUGIN_INSTALL_DIR" ]; then
-    rm -r $PLUGIN_INSTALL_DIR  
-	echo "[INFO] Have deleted the $PLUGIN_INSTALL_DIR"
+# Interaction for sure to detele files
+read -n 1 -p "Are you sure to delete the ~/.vim and the ~/.vimrc [Y/N]?" ANSWER
+echo ""
+if [ $ANSWER = 'Y' -o  $ANSWER = 'y' ]; then
+   #delete .vim
+    if [ -d "$PLUGIN_INSTALL_DIR" ]; then
+        rm -r $PLUGIN_INSTALL_DIR
+        echo "[INFO] Have deleted the $PLUGIN_INSTALL_DIR"
+    else
+        echo "[WARNING] Have no $PLUGIN_INSTALL_DIR"
+    fi
+
+    #delete .vimrc
+    if [ -f "$CONFIG_INSTALL_DIR" ]; then
+        rm $CONFIG_INSTALL_DIR
+        echo "[INFO] Have deleted the $CONFIG_INSTALL_DIR"
+    else
+        echo "[WARNING] Have no $CONFIG_INSTALL_DIR"
+    fi
+
+    echo "Uninstall finished!"
+else
+    echo "Do nothing!"
+    exit 1
 fi
-
-#delete .vimrc
-if [ -f "$CONFIG_INSTALL_DIR" ]; then
-    rm $CONFIG_INSTALL_DIR  
-	echo "[INFO] Have deleted the $CONFIG_INSTALL_DIR"
-fi
-
-#delete .ycm_extra_conf
-if [ -f "$YCM_CONFIG_INSTALL_DIR" ]; then
-    rm  $YCM_CONFIG_INSTALL_DIR  
-	echo "[INFO] Have deleted the $YCM_CONFIG_INSTALL_DIR"
-fi
-
-echo "Uninstall finished!"
